compressed species files
These are the XML code segments for each species in my data source. Any combination of the species with a "beginning.kml" and an "end.kml" would make a functional kml file that can be displayed on Google Earth. The precision of the geospatial data used is about 1 meter on Earth.